split view and controller into seprate files and update view controller
