+++ app/src/modules/components/character/characterController.js
@@ -1,0 +1,1 @@
+module.exports = { }
