+++ app/src/modules/components/character/characterView.js
@@ -1,0 +1,13 @@
+module.exports = function() {
+    const view = `
+    <div id="character-view">
+    
+        <div id="cv-header">
+            <h2>Character</h2>
+        </div>
+    
+    </div>
+    `;
+    
+    return view;
+}
