--- app/src/modules/components/characterComponent.js
@@ -1,16 +1,0 @@
-module.exports = {
-    viewTemplate: function() {
-        const view = `
-        <div id="character-view">
-        
-            <div id="cv-header">
-                <h2>Character</h2>
-            </div>
-        
-        </div>
-        `;
-        
-        return view;
-    },
-    viewController: { }
-}
