+++ app/src/modules/components/dungeon/dungeonController.js
@@ -1,36 +1,9 @@
 // Required Modules
-const state = require('../state/stateController');
-const combat = require('../combat/combatcontroller');
+const state = require('../../state/stateController');
+const combat = require('../../combat/combatcontroller');
 
 // Export of module object
 module.exports = {
-    // View method that returns the view template string 
-    viewTemplate: function() {
-        const view = `
-        <div id="dungeon-view">
-        
-            <div id="dv-header">
-                <h2>Dungeon</h2>
-            </div>
-        
-            <div id="rendered-dungeon">
-        
-                <button id="enter-dungeon-btn">Enter Dungeon</button>
-        
-                <div id="dungeon-messages"></div>
-        
-                <div id="entity-cards"></div>
-        
-                <div id="action-btns"></div>    
-            </div>
-        
-        </div>
-        `;
-
-        return view;
-    },
-    // Controller for the logic of dungeon view and view rendering
-    dungeonController: {
         // Event Listeners
         loadListeners: function() {
             // load Event Listeners
@@ -99,5 +72,3 @@ module.exports = {
             });
         }
     }
-
-}
